% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jsdiff}
\alias{jsdiff}
\title{Create a diff view using jsdiff}
\usage{
jsdiff(oldStr, newStr, diffType = c("Chars", "Words", "Lines"),
  language = "r", width = NULL, height = NULL)
}
\arguments{
\item{oldStr}{Old string of text to diff}

\item{newStr}{New string of text to diff}

\item{diffType}{Type of diff to use}

\item{language}{The syntax highlighting language to use}

\item{width}{Fixed width for widget (in css units). The default is NULL, which results in intelligent automatic sizing based on the widget's container.}

\item{height}{Fixed height for widget (in css units). The default is NULL, which results in intelligent automatic sizing based on the widget's container.}
}
\description{
This creates a jsdiff widget which allows user to see differences between two strings.
}

