% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jsdiff.R
\name{jsdiffOutput}
\alias{jsdiffOutput}
\title{Widget output function for use in Shiny}
\usage{
jsdiffOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{of the widget element}

\item{height}{of the widget element}
}
\description{
Widget output function for use in Shiny
}

